+++ packages/flutter_deck/example/lib/slides/end_slide.dart
@@ -1,12 +1,17 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_deck/flutter_deck.dart';
 
+const _speakerNotes = '''
+- Use flutter_deck to create your own slides.
+''';
+
 class EndSlide extends FlutterDeckSlideWidget {
   const EndSlide()
       : super(
           configuration: const FlutterDeckSlideConfiguration(
             route: '/end',
             title: 'Thank you!',
+            speakerNotes: _speakerNotes,
             footer: FlutterDeckFooterConfiguration(showFooter: false),
           ),
         );